backend/requirements.txt
fastapi
uvicorn
google-adk==1.10.0
python-dotenv
pydantic
google-generativeai